Fix for bottom button
Bottom button which helps user get back to top of the page were not
working because it were linking to #myNavmenu which were not existing on
page because of display: none property. This property is set by external
library and is needed.
I added #top to .jubmotron and linked button to it.